refactor(frontend): move workspace provider into workspaces presentation layer
