+++ frontend/src/modules/workspaces/presentation/index.ts
@@ -1,1 +1,1 @@
-export {};
+export * from "./workspace-provider";
